HDRenderPipeline: split class and make it work
