+++ Assets/ScriptableRenderPipeline/HDRenderPipeline/Lighting/AmbientOcclusion/ScreenSpaceAmbientOcclusionSettingsSingleton.cs
@@ -1,0 +1,13 @@
+namespace UnityEngine.Experimental.Rendering.HDPipeline
+{
+    public class ScreenSpaceAmbientOcclusionSettingsSingleton : Singleton<ScreenSpaceAmbientOcclusionSettingsSingleton>
+    {
+        private ScreenSpaceAmbientOcclusionSettings settings { get; set; }
+
+        public static ScreenSpaceAmbientOcclusionSettings overrideSettings
+        {
+            get { return instance.settings; }
+            set { instance.settings = value; }
+        }
+    }
+}
