+++ Assets/ScriptableRenderPipeline/HDRenderPipeline/Lighting/AmbientOcclusion/ScreenSpaceAmbientOcclusionSettingsSingleton.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 5da3bcf229fbf3d4ab232413e3a07457
+timeCreated: 1496196658
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
